UI.h: Oops. Forgot to cast len back to int when assigning to caret.

+++ UI.h
@@ -331,7 +331,7 @@ namespace ui
     {
       auto len = std::min(input.length(), str.length());
       input.replace(0, len, str.substr(0, len));
-      caret = len;
+      caret = static_cast<int>(len);
     }
     
     void clear_input()
